requirements.txt
anthropic>=0.40.0
PyGithub>=2.1.1
pyyaml>=6.0
